integration: Refactor tar and dockermode together
